# cuda4dnn: overlap D2H output blobs transfer with inference